requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
bitarray==2.7.3
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
channels==3.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==40.0.2
cytoolz==0.12.1
daphne==3.0.2
DateTime==5.1
Django==4.2.1
django-admin-material-dashboard==1.0.10
django-colorfield==0.8.0
djangorestframework==3.14.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
gunicorn==20.1.0
hexbytes==0.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonschema==4.17.3
kombu==5.2.4
lru-dict==1.1.8
msgpack==1.0.5
multidict==6.0.4
parsimonious==0.9.0
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.31.0
rlp==3.0.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toolz==0.12.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
web3==6.4.0
websockets==11.0.3
whitenoise==6.4.0
yarl==1.9.2
zope.interface==6.0
